sylli_crawl->driver_manager.py: move selenium driver to stand alone module
This commit moves the selenium driver out of the js crawler and in
to a stand alone module.
